SampleOn Mouse.isDown Mouse.position
Using this is the most reliable measure of mouse position and user
activities as we know when the user clicks down and releases the mouse
button. This way we can model selecting, dragging, and it does not
interfere with the codemirror.js binding to mouse clicks, so the editor
is unaffected! Brilliant!